Addressed comments

+++ internal/venice-common/src/main/java/com/linkedin/venice/common/Measurable.java
@@ -1,5 +1,5 @@
 package com.linkedin.venice.common;
 
 public interface Measurable {
-  int getSize();
+  long getSize();
 }
